Add total length to record structure
This is necessary for record trees. While it could be added separately
reserving space inside the record is easier & more efficient.

The reference count has been reduced from 64 to 8 bits. This should
still be plenty in practice. In the worst case, it won't be able to
deduplicate a file more than 254 times.