Replace 32-byte arbitrary hash with 8-byte XXH3
For error detection a cryptographic hash is overkill and for
deduplication it is more efficient to first look at the length and only
then compare file contents.

While hashing would speed up the latter process it actually doesn't
reduce the real cost. The cost is merely paid at the time of writing
instead of the time of reading. Hashes can also be cached by an external
tool, further reducing the need to create a hash at the time of writing.