Remove Dir::move_child, simplify HashMap::insert
The latter was too easily misused and is removed in favor of doing the
equivalent operations ad-hoc.

In HashMap::insert it is only necessary to determine the insertion point
based on PSL for the entry to be inserted.
The other entries are already in order and can simply be shifted
forward.