Various dir fixes + discovery that readdir() is broken
Specifically, after much thought, it is *our* implementation that is
broken.

Suppose we want to delete a lot of files one-by-one (i.e. how it is done
on every major OS).
For it to work with NRFS on Linux as of writing we need to:

* opendir()
* readdir()
* unlink()
* closedir()

That's going to be *very* slow when the first half of the hashmap turns
into empty entries...

Even more: readdir() ought to work for any files that have not been
added or removed after the opendir() call, so our implementation is
broken regardless.

The specification will need some amendments to make it workable.
We can find a proper solution for the "dangling inode" problem as well
then.