Rework dir
The old directory structure consisted of a hashmap, which contained item
data directly, and a separate heap.
This turns out to not work properly with readdir(), especially in an
efficient (i.e. less than O(n^2)) manner when removing entries.

To fix this, the directory is now split in three parts:

* A list of items.
* A hashmap whose entries point to these items.
* A heap.

This allows iteration over a directory with readdir() while adding and
removing entries without affecting current entries.