Implement write_zeros
After scribbling on many pages of paper I figured out that Tree::shrink
could be simplified by separating some of the functionality in a
separate function: write_zeros.
This function will also be useful if a user needs to clear a large
region quickly for whatever reason.

While sketching out how to implement write_zeros nicely I made a second
realization: It is possible to zero out records efficiently & bottom-up
by looking for records that are already zero and going up & down the
tree as appropriate - the higher on the tree, the quicker iteration
goes.

The current implementation is very hacky and some existing methods have
temporarily been stubbed with todo!(). The next commit will clean it up.