Rewrite Tree::shrink to use write_zeros
The main difficulty with Tree::shrink was efficiently removing a
potentially large range of records.
Since replacing non-zero records with zero records is effectively the
same as removing the record write_zeros is practical to use for this
purpose.

The Tree::shrink function is now simple to comprehend and actually
works, at long last!