Add callgrind script
As it turns out, an absolutely *massive* amount of time is spent just
copying data: 30%!

The second worst offender is _int_malloc and _int_free, for a total of
10.5 + 8.5 = 19% of time.

The third worst offender is "<cycle 2>", where about 10% of time is
spent. I'll need a mapping to source code to figure out what exactly it
is though.

The remainder is not worth looking at too much at the moment, since
almost half of time is spent on the three functions above.