Fix stale fetches after shrink
This was a gigantic pain in the ass to resolve, but it should be
correct at long last.

The main new mechanic here is "chain hopping" in Tree::get.

If a fetch for an entry is already in progress, e.g. due to
evict_entry, resizing the object may interfere with it by
"splitting" the Tree::get chain.

For example, suppose this chain is being fetched from top
to bottom:

        |1|       d2
        /
      |1|         d1
      /
    |1|           d0

If a resize happens, the chain may look like this:

        |2|       d2
        /
      |1|         d1
      /
    |1|           d0

Tree::get must now "hop" from 2 to 1.
However, it won't find the correct record at d2, since it has been
moved out of the entry to an object's root.
Tree::get must instead get that object's root to continue.

While at it, fetch_entry has been removed since it has been entirely
substituted by Tree::get and Tree::set.
The latter can directly insert a record without redundantly fetching it
if all data would be overwritten anyways.
In all other cases (except write_zeros) Tree::get is necessary.