Manually parallelize fuzzer
For whatever reason libFuzzer's fork mode constantly spawns new
processes whereas doing it manually keeps each process alive.

This probably has a fair bit of overhead so it is being done manually
again.

Not to mention, the output of fork mode is quite confusing: is it really
only executing at the speed of a single worker?