Rewrite Tree::get
The old implementation wasn't able to handle multiple grows/shrinks.

To work around it it is necessary to fetch the proper record again if
the ID of the child and its supposed parent differ, i.e. a "chain break"
is detected, i.e. in pseudo-code:

    get
      if target entry present
        return target entry
      loop
        get parent entry
        if chain break
          retry loop
	return target entry

There are numerous other smaller changes which have been added to make
things simpler to use & debug.