Set block size to 12 by default, limit range to 9..=24
It is better to default to a block size that may be larger than strictly
required to account for devices that may be added to the pool later.

e.g. one may create a pool with a device that has a sector size of 512,
then later add a device with a sector size of 4096.

The range of record size and block size have been limited to 9..=24
since any values outside of it are invalid anyways.