WIP: hard memory limit, remove Background from public API
Turns out all we needed to properly abstract Background was just a
separate task running concurrently. How convenient.

The hard memory limit is introduced to prevent huge memory peaks
(several times the soft limit!).
If it would be exceeded tasks are forced to yield.

It isn't complete yet, as EntryRef::{write,replace} are still able to
exceed it.
It should be possible to hard limit them properly, but I haven't figured
out a clean way to do it yet.

A bug has also been fixed in the process of introducing the hard limit:
Bitmap index calculation caused past-of-record writes.
Alas, this means that any filesystem with more than 4 * max_record_size
of objects is in an inconsistent state :(