Simplify cache memory tracking
This design assumes every entry has the same size regardless of real
size. This simplifies the implementation significantly as there is no
need to account for resizes.

A major potential use is the under-utilization of the cache, which could
lead to poor performance for many small random I/O operations. In
practice, this _probably_ won't be an issue since:

* Ideally most operations hit the VFS cache only, leaving the object
  store out of the picture.
* The device drivers likely uses buffers, which may also act as a sort
  of cache.

Overall, I expect that the performance will be reasonable even in
regressive cases.