nrfs: Respecify directory entry, reintroduce multi-object dir
Using a single object proved to be a fair bit too annoying, so
multi-object dirs have been reintroduced.

Items are now variable sized. This allows avoiding needing to put long
names on the heap, further reducing complexity.
It is designed such that readdir() will never start reading from the
middle of an item, even with a malicious seekdir() call.